[New] support new config system
  - add configs/ entry points for recommended configs
  - add documentation to readme
